str.delete!("=") iterates over the entire string looking for the equals
character, but we know that we will, at most, find 2 at the end of the
string.