@jeremyevans 2021-09-02
Can you please submit this upstream to https://github.com/ruby/base64?