@jcmfernandes 2021-09-02
Some benchmarks:

Input generated with `SecureRandom.alphanumeric`.
`.urlsafe_encode64` being called with `padding: false`.

```
4 kB input

Warming up --------------------------------------
NEW #urlsafe_encode64
                        14.230k i/100ms
OLD #urlsafe_encode64
                         7.588k i/100ms
Calculating -------------------------------------
NEW #urlsafe_encode64
                        141.104k (± 2.5%) i/s -      1.423M in  10.091936s
OLD #urlsafe_encode64
                         88.993k (± 5.5%) i/s -    887.796k in  10.009833s

Comparison:
NEW #urlsafe_encode64:   141104.0 i/s
OLD #urlsafe_encode64:    88993.5 i/s - 1.59x  (± 0.00) slower
```

```
4 MB input

Warming up --------------------------------------
NEW #urlsafe_encode64
                        13.000  i/100ms
OLD #urlsafe_encode64
                         8.000  i/100ms
Calculating -------------------------------------
NEW #urlsafe_encode64
                        136.878  (± 2.9%) i/s -      1.378k in  10.074976s
OLD #urlsafe_encode64
                         87.759  (± 2.3%) i/s -    880.000  in  10.030723s

Comparison:
NEW #urlsafe_encode64:      136.9 i/s
OLD #urlsafe_encode64:       87.8 i/s - 1.56x  (± 0.00) slower
```